requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.7
django-ckeditor==6.1.0
django-js-asset==1.2.2
filelock==3.0.12
Pillow==8.3.2
platformdirs==2.3.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.8.0
virtualenvwrapper-win==1.2.6
